✨ Sale service.

+++ src/main/java/henrotaym/env/annotations/ExistsInDatabase.java
@@ -7,7 +7,6 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Constraint(validatedBy = ExistsInDatabaseValidator.class)
@@ -22,5 +21,5 @@ public @interface ExistsInDatabase {
 
   Class<? extends Payload>[] payload() default {};
 
-  Class<? extends JpaRepository<?, BigInteger>> repository();
+  Class<? extends JpaRepository<?, Long>> repository();
 }
