+++ src/main/java/henrotaym/env/validators/ExistsInDatabaseValidator.java
@@ -3,15 +3,13 @@ package henrotaym.env.validators;
 import henrotaym.env.annotations.ExistsInDatabase;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import java.math.BigInteger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public class ExistsInDatabaseValidator
-    implements ConstraintValidator<ExistsInDatabase, BigInteger> {
+public class ExistsInDatabaseValidator implements ConstraintValidator<ExistsInDatabase, Long> {
 
-  private Class<? extends JpaRepository<?, BigInteger>> repositoryClass;
+  private Class<? extends JpaRepository<?, Long>> repositoryClass;
 
   @Autowired ApplicationContext applicationContext;
 
@@ -21,7 +19,7 @@ public class ExistsInDatabaseValidator
   }
 
   @Override
-  public boolean isValid(BigInteger value, ConstraintValidatorContext context) {
+  public boolean isValid(Long value, ConstraintValidatorContext context) {
     if (value == null) {
       return false;
     }
